Views の命名をパスカルに / 構造修正
