+++ src/views/Manages/Favorites/Favorite.vue
