+++ src/views/Users/Registers/RegisterStep2.vue
