+++ src/views/Users/Registers/RegisterStep3.vue
