+++ src/views/Users/Registers/SentMailComplete.vue
